+++ packages/chat/src/components/Carousel/hooks.ts
@@ -1,7 +1,7 @@
 import type { RefObject } from 'react';
 import { useEffect, useRef, useState } from 'react';
 
-import type { CardProps } from '../Card';
+import type { CardProps } from '../Card/types';
 import { CARD_WITH_GUTTER_WIDTH, NEXT_CONTROL_BOUNDARY, PREVIOUS_CONTROL_BOUNDARY } from './constants';
 import { CAROUSEL_GUTTER_WIDTH } from './styled';