+++ packages/chat/src/stories.ts
@@ -1,10 +1,12 @@
 export * as AgentMessage from './components/AgentMessage/AgentMessage.story';
 export * as Avatar from './components/Avatar/Avatar.story';
 export * as Button from './components/Button/Button.story';
+export * as Card from './components/Card/Card.story';
 export * as ColorTheme from './components/ColorTheme/ColorTheme.story';
 export * as FeedbackButton from './components/FeedbackButton/FeedbackButton.story';
 export * as Header from './components/Header/Header.story';
 export * as Launcher from './components/Launcher/Launcher.story';
+export * as LinkPreview from './components/LinkPreview/LinkPreview.story';
 export * as MessageInput from './components/MessageInput/MessageInput.story';
 export * as SendButton from './components/SendButton/SendButton.story';
 export * as UserMessage from './components/UserMessage/UserMessage.story';